refactored create-validator
