+++ jest-no-db.config.ts
@@ -6,5 +6,6 @@ export default {
   testMatch: [
     "<rootDir>/src/__test__/controllers/**",
     "<rootDir>/src/__test__/middlewares/**",
+    "<rootDir>/src/__test__/schemas/**",
   ],
 };
